+++ tests/test_models.py
@@ -65,6 +65,22 @@ class TestRegistryResponse:
         assert rr.count == 0
         assert len(rr.applications) == 0
 
+    def test_load_from_json_string_applications_none(self):
+        data = {
+            'query_uri': 'http://id.example.org/me',
+            'success': True,
+            'has_references': False,
+            'count': 0,
+            'applications': None
+        }
+        data = json.dumps(data)
+        rr = RegistryResponse.load_from_json(data)
+        assert rr.query_uri == 'http://id.example.org/me'
+        assert rr.success
+        assert not rr.has_references
+        assert rr.count == 0
+        assert len(rr.applications) == 0
+
 
 class TestApplicationResponse:
 
@@ -129,6 +145,26 @@ class TestApplicationResponse:
         assert ar.has_references
         assert ar.count == 1
         assert len(ar.items) == 1
+
+    def test_load_from_json_string_no_items(self):
+        data = {
+            'title': 'Zotskapp',
+            'uri': 'http://zotskapp.be',
+            'service_url': 'http://zotskapp.be/references',
+            'success': True,
+            'has_references': False,
+            'count': 0,
+            'items': []
+        }
+        data = json.dumps(data)
+        ar = ApplicationResponse.load_from_json(data)
+        assert isinstance(ar, ApplicationResponse)
+        assert ar.uri == 'http://zotskapp.be'
+        assert ar.service_url == 'http://zotskapp.be/references'
+        assert ar.success
+        assert not ar.has_references
+        assert ar.count == 0
+        assert len(ar.items) == 0
 
 class TestItem: